requirements.txt
flask
pyyaml
requests
transformers
torch
flask-cors